refactor(chart): share sparkline y-scale and baseline drawing
drawMiniLine and drawMiniCandles had byte-identical y-scale math
(padY/innerH/range/yOf, folding the baseline into the extent) and an
identical dashed-baseline block. Extract computeYScale() and
drawSparkBaseline() into sparkline/scale.ts and call them from both.
Each renderer keeps its own raw min/max scan (line: close; candle:
low/high). Kept separate from the main-chart series renderers on
purpose so the PriceScale/TimeScale machinery stays out of the
sparkline bundle. size-check: Sparkline (vanilla) 3.78 kB / 5 kB.
